Refactor integration modal with tabs
